dentistav1/requirements.txt
Flask==2.0.3
google-cloud-vision==3.0.0
openai==1.0.0
python-dotenv==1.0.1
gunicorn==20.1.0
